refactor(settings): separate scroll policy and view adapters
