+++ Packages/macOS/CmuxSettingsUI/Sources/CmuxSettingsUI/Navigation/SettingsSectionScrollGeometry.swift
@@ -1,24 +1,13 @@
 import CoreGraphics
-import SwiftUI
 
+/// Combines header and content bounds measured in the same viewport coordinate space.
 struct SettingsSectionScrollGeometry: Equatable, Sendable {
     var positions: [SettingsSectionScrollPosition] = []
     var contentBottomY: CGFloat?
 
+    /// Excludes trailing padding and remains invariant as the viewport scrolls.
     var lastSectionHeight: CGFloat? {
         guard let contentBottomY, let lastHeaderY = positions.map(\.minY).max() else { return nil }
         return max(0, contentBottomY - lastHeaderY)
-    }
-}
-
-struct SettingsSectionScrollGeometryKey: PreferenceKey {
-    static let defaultValue = SettingsSectionScrollGeometry()
-
-    static func reduce(value: inout SettingsSectionScrollGeometry, nextValue: () -> SettingsSectionScrollGeometry) {
-        let next = nextValue()
-        value.positions.append(contentsOf: next.positions)
-        if let contentBottomY = next.contentBottomY {
-            value.contentBottomY = contentBottomY
-        }
     }
 }
