+++ Packages/macOS/CmuxSettingsUI/Sources/CmuxSettingsUI/Navigation/SettingsSectionScrollPosition.swift
@@ -1,0 +1,7 @@
+import CoreGraphics
+
+/// A section header's vertical position in the settings detail viewport.
+struct SettingsSectionScrollPosition: Equatable, Sendable {
+    let section: SettingsSectionID
+    let minY: CGFloat
+}
